+++ CHANGES.rst
@@ -15,6 +15,7 @@ Improvements:
  * MXStore: Add a method to get related events for a specific event.
  * MXPublicRoom: Add canonical alias property.
  * MXLogger: Add a parameter to indicate the number of log files.
+ * MXDeviceList: Post `MXDeviceListDidUpdateUsersDevicesNotification` notification when users devices list are updated.
 
 Bug fix:
  * MXEventType: Fix Swift refinement.